+++ packages/nutui/components/progress/types.ts
@@ -1,0 +1,2 @@
+export const progressStatus = ['text', 'active', 'icon'] as const
+export type ProgressStatus = (typeof progressStatus)[number]